Move the conditions tool into src/components/conditions/
Twenty-two modules and 41 files — just under half the directory — that
had no name. Of their outbound relative imports only four leave the
subsystem, and all four land in `ui/`, so the whole tool is one directory
with a four-edge boundary.

The eight `vi.mock` strings are the part the type checker cannot see. All
eight are updated here, and the assertions around four of them read text
only the mock factory produces — "panel for <slug>", "waves for", "wind
for", "week for" — so a mock that quietly stopped applying fails the
suite rather than passing on the real module.

The four helpers come too. `cardText`, `disclosure`, `headingRank` and
`mopLine` are each read only by panels in this directory; `disclosure`'s
reach into `ui/touchTarget` is one of the four boundary edges.
