+++ src/components/conditions/Caveats.tsx
