+++ src/components/conditions/ConditionsNotes.tsx
