+++ src/components/conditions/ReadingCard.tsx
