+++ src/components/conditions/StatGroup.tsx
