+++ src/components/conditions/WavePanel.test.tsx
