+++ src/components/conditions/WavePanel.tsx
