+++ src/components/conditions/WeekPanel.tsx
