+++ src/components/conditions/WindPanel.test.tsx
