+++ src/components/conditions/WindPanel.tsx
