+++ src/components/conditions/cardText.ts
