+++ src/components/conditions/headingRank.ts
